block queue remoted
