+++ webcam-capture-drivers/driver-raspberrypi/src/main/java/com/github/sarxos/webcam/ds/raspberrypi/RaspistillDevice.java
@@ -4,9 +4,6 @@ import java.awt.image.BufferedImage;
 import java.io.IOException;
 import java.util.Map;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 /**
  * ClassName: RaspistillDevice <br/>
  * date: Jan 31, 2019 12:16:12 PM <br/>
@@ -16,7 +13,6 @@ import org.slf4j.LoggerFactory;
  * @since JDK 1.8
  */
 public class RaspistillDevice extends IPCDevice {
-	private final static Logger LOGGER = LoggerFactory.getLogger(RaspistillDevice.class);
 	/**
 	 * raspistill keypress mode, send new line to make capture
 	 */
@@ -50,10 +46,6 @@ public class RaspistillDevice extends IPCDevice {
 		return new ErrorConsumeWorker();
 	}
 	
-	@Override
-	protected Runnable newCaptureWorker() {
-		return new CaptureWorker();
-	}
 	/**
 	 * One github story: I interchanged the order of "-t 0" and "-s" and tested it. Code: Select all
 	 * raspistill -t 0 -s -o test.jpg And signal driven event works now as expected!
@@ -70,44 +62,17 @@ public class RaspistillDevice extends IPCDevice {
 		parameters.put(OPT_OUTPUT, "-");// must be this, then image will be in console!
 	}
 
-	// inner classes
-	class CaptureWorker implements Runnable {
-		@Override
-		public void run() {
-			while (!Thread.currentThread().isInterrupted()) {
-				try {
-					out.write(CAPTRE_TRIGGER_INPUT);
-					out.flush();
-
-					BufferedImage frame = new PNGDecoder(in).decode();
-					in.skip(16);// each time of decode, there will be 16 bytes should be skipped
-					frameBuffer.offer(frame);
-				} catch (IOException e) {
-					LOGGER.info(e.toString());
-				}
-			}
+	@Override
+	public BufferedImage getImage() {
+		try {
+			out.write(CAPTRE_TRIGGER_INPUT);
+			out.flush();
+			BufferedImage frame = new PNGDecoder(in).decode();
+			in.skip(16);// each time of decode, there will be 16 bytes should be skipped
+			return frame;
+		} catch (IOException e) {
+			e.printStackTrace();
 		}
-	}
-
-	class ErrorConsumeWorker implements Runnable {
-		@Override
-		public void run() {
-			while (!Thread.currentThread().isInterrupted()) {
-				try {
-					Thread.sleep(3000);
-				} catch (InterruptedException e) {
-					LOGGER.info(e.toString(), e);
-					break;
-				}
-				try {
-					int ret = -1;
-					do {
-						ret = err.read();
-					} while (ret != -1);
-				} catch (IOException e) {
-					LOGGER.info(e.toString(), e);
-				}
-			}
-		}
+		return null;
 	}
 }
